# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "LaudryApp",
    "description": "",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "schemes": [],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/register": {
      "post": {
        "description": "",
        "summary": "Register",
        "operationId": "Register",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "nama",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "telepon",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/x-www-form-urlencoded"
            ],
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/login": {
      "post": {
        "description": "",
        "summary": "Login",
        "operationId": "Login",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/x-www-form-urlencoded"
            ],
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/layanan": {
      "post": {
        "description": "",
        "summary": "Layanan",
        "operationId": "Layanan",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "nama",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "unit",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "harga",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/x-www-form-urlencoded"
            ],
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "security": [],
  "tags": []
}